falter-berlin-network-defaults: add no-internet-ssid-changer
Activate panic.d scripts and set the timer to 1 minute.
If the node is not connected to the internet longer for 1 minute the
ssid will change to 'no-internet.berlin.freifunk.net'.

If the node is back online the ssid will change automatically back to
the default freifunk ssid.

Requested by #4.

Signed-off-by: Nick Hainke <vincent@systemli.org>