falter-berlin-migration: handle profiles
Upon upgrade, the files /etc/config/profile_* and /etc/config/freifunk
are not automatically upgraded to the newest version.  This commit
pulls the profile_* out of the rom partition and overwrites the
profiles in /etc/config.

Additionally the freifunk config is handeled seperately.  This config
file is a mixure of defaults with the addition of the 'contact' and
'community' sections which are configured for the running system.
In this case, the version from rom is copied into /etc/config and
the 'contact' and 'community' sections from the old setup
are imported.

this is based on the work in https://github.com/freifunk-berlin/firmware-packages/commit/c9c363bd19f79e30b192109d916d81fc0d5f41d4

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>