luci-app-ffwizard-falter: add upgrade wizard support
The ability to add upgrades, which the user would need to confirm upon
logging in after running a firmware upgrade, are added with this commit.

All files which are located in /usb/lib/lua/luci/model/cbi/freifunk/upgrades
are added to the list of upgrades to be done.  The filenames
should have the format NNN-upgradeDescription.lua

The order of the upgrades are important, and new upgrades should
have a higher NNN number at the beginning.  In order not to repeat
an upgrade page, the value ffwizard.upgrade.lastUpgrade is used.

Each of these lua files need to:
* set the value ffwizard.upgrade.lastUpgrade to the filename (without ".lua")
* Upon Apply, do a redirect back to startWizard, which will either
forward the user to the next upgrade page or to the default status page.

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>