falter-berlin-migration: give wireless sections names
In OpenWrt 19.07 the wifi-iface sections are no longer
nameless.  This migration script will set the names to
the same as the ifname (with - substituted with _).  If
an wifi-iface section already has a name, nothing happens.
If an unnames wifi-iface doesn't have the ifname property
set, then wifinetN is used, where N is the number of the
wifi-iface.

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>