luci-app-ffwizard-falter: check mesh modes in wizard
Add the ability to check the current driver to see which modes
"802.11s" or "Ad-hoc" are supported.  Only give the user the option
to select a mode if it is a valid mode.  This can be seen on
both the wireless wizard page and on the wireless mesh upgrade
page.

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>