falter-berlin-migration: handle sharenet
when ffwizard.settings.sharenet is not set to 1 (don't share),
then set the ffuplink interface to disabled=1.  Otherwise the
policy routing will not provide internet (desired) but still
olsrd will advertise and up (black hole/not desired).

if sharing is desired (set to 1), then explicitly set the ffupink
to disabled=0

Fixes: #126
Signed-off-by: pmelange <isprotejesvalkata@gmail.com>