luci-mod-falter: add BBB-VPN (bbbdigger) config page
Add the ability to enable (create if not already created) or
disable the bbbdigger interface.  If the interface and related
settings have not been created before, use the default values
hard coded into the page.  If the interface and related settings
have been created before, simple toggle between disabled=1 and
disabled=0 (or enabled/ignore) in config files network, olsrd
and tunneldigger as needed.

This is a very Berlin centric page and is not easily portable
to other communities.  If portablility to other communitites is
required, then a rework of the settings should be considered.

Fixes: #54
Signed-off-by: pmelange <isprotejesvalkata@gmail.com>