falter-common: rework freifunk_release
Change the meaning of the variable FREIFUNK_OPENWRT_BASE to
reflect the SDK and Imagebuilder found on downloads.openwrt.org.
This variable points to the location in the "releases" directoy
except when FREIFUNK_OPENWRT_BASE is set to "master".  In that
case, the variable points to the snapshot builds in the
"snapshots" directory.

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>