luci-app-falter-owm: reimplement owm.lua in sh
As OpenWrt will stop its use of lua in the next time, we decided
to reimplement the owm.lua script, which pushes node data to the
openwifimap api. This script implements some kind of owm-api v1.0.

It's a basis for the new owm-api-script, which we can write, after
we found a definition for the new format.

Signed-off-by: Martin Hübner <martin.hubner@web.de>