falter-berlin-network-defaults: defaults for openwrt 21.02
adjust defaults for compatability with openwrt 21.02.  This
change is a bit more difficult because of 3 factors.

1) if the wan port does not exist (single port device), create
a wan device and wan/wan6 interfaces which do not have physical
ports.  This makes it easier to change the single physical port
from the client network to wan at a later time
2) if the wan interface is already part of a bridge, which is what
we want, then do nothing special
3) if the wan interface is only a single device port, then change
it to a bridge device

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>