falter-berlin-tunneldigger: add SRV nslookup support
Add support for option srv, which denotes an SRV
DNS entry.  This can be used instead or with the address list.
The advandate of using an srv, is that only on the dns server
side do we need to modify the list as servers are added/removed.

Fixes: #220
Signed-off-by: pmelange <isprotejesvalkata@gmail.com>
(cherry picked from commit 29943b0d10ac3e04322dbe222e0c6453a37b49e3)