falter-berlin-tunnelmanager: several fixes (#278)
* falter-berlin-tunnelmanager: don't try to connect to same server on error

If a server denies a request, tunnelmanager ignores it and always tries
to connect to it again, if it is the server with the least users.

In case a server denies a request, search again for another server.

Signed-off-by: Nick Hainke <vincent@systemli.org>

* falter-berlin-tunnelmanager: replace vpnmanager with tunnelman

Signed-off-by: Nick Hainke <vincent@systemli.org>

* falter-berlin-tunnelmanager: switch to ipc in up/down

Switch to using ipc for the up/down scripts. Adding and removing will be
seamless.

Signed-off-by: Nick Hainke <vincent@systemli.org>

* falter-berlin-tunnelmanager: cleanup interface config before adding new interfaces

Signed-off-by: Nick Hainke <vincent@systemli.org>

* falter-berlin-tunnelmanager: cleanup idents

Use
    shfmt -i 4 -w $FILE
to use spaces instead of tabs.

Signed-off-by: Nick Hainke <vincent@systemli.org>

* falter-berlin-tunnelmanager: use "=" for checking strings

"==" is not defined. Use "=" instead.

Signed-off-by: Nick Hainke <vincent@systemli.org>

* falter-berlin-tunnelmanager: increment version

Signed-off-by: Nick Hainke <vincent@systemli.org>