falter-berlin-tunnelmanager: increase timeout to register interface
Should fix #280.

Should fix #280.

5 seconds seem to be to fast to register an interface. I suspect that
the client registers some interface, a timeout triggers, and the server
installed the interface. However, due to the timeout, the client excepts
that a failure occurs. That continues in a loop and will prevent a
client to connect to some gateway again. Increase timeout to 1 minute to
allow the client to have enough time to install the wireguard interface.

Signed-off-by: Nick Hainke <vincent@systemli.org>