luci-app-ffwizard-falter: fix nil-error
On some device we arbitrarily experienced some issues with getting nil-
values from a rpc-call. This call figures out if there was a root password
set already. As nil-values crash the LuCI-Interface, we provide an
alternative value if the call returns nil, which is assuming that the
password was not set.

Fixes #298

Signed-off-by: Martin Hübner <martin.hubner@web.de>