tunnelmanager: start it earlier
We want a working tunnel as fast as possible, but 95 is pretty late
in the boot process. network is at 20 and olsrd/babel at 65/70,
so 40 should be fine.

Also stop a little later - tunneldigger doesn't need to be among the
first services that are stopped.

Signed-off-by: Packet Please <pktpls@systemli.org>
(cherry picked from commit 682de59c4bd58c0c84f471d1767143e7ddb2b29a)