falter-berlin-ssid-changer: several fixes
In the merged state before, the ssid-changer was only functional, if there
did happen a change in the state of interfaces. But if a router started
without having internet connectivity, it wasn't working at all.

This commit adds a check for connectivity running every 5 minutes, hopefully
increasing the user experience a lot.

Signed-off-by: Martin Hübner <martin.hubner@web.de>